pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xautomation"
version = "0.1.0"
description = "X timeline scraper with rule + optional LLM filtering"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
xauto = "xautomation.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
